Add files via upload
These are the Jane Porter files. This program creates a simple python server that opens the port of your choice and allows for two way file transfer. It makes files in the directory that it is opened in available for download by whomever logs into the IP:PORT website, and also upload files to the IP:PORT, after the host grants permission, per file. This version only allows for transference of one file at a time.